🚚 rename: App.js, Cars.js app 폴더 내부로위치 이동
